--- a/apps/design/backend/src/types.ts
+++ b/apps/design/backend/src/types.ts
@@ -1,4 +1,4 @@
-import z from 'zod';
+import z from 'zod/v4';
 import {
   BallotStyle as VxfBallotStyle,
   BallotStyleId,