--- a/apps/mark-scan/backend/src/globals.ts
+++ b/apps/mark-scan/backend/src/globals.ts
@@ -1,7 +1,7 @@
 /* istanbul ignore file - @preserve */
 import { unsafeParse } from '@votingworks/types';
 import { join } from 'node:path';
-import { z } from 'zod';
+import { z } from 'zod/v4';
 
 /**
  * Default port for the VxMarkScan API.