--- a/libs/api/src/services/print-scan/index.ts
+++ b/libs/api/src/services/print-scan/index.ts
@@ -1,4 +1,4 @@
-import { z } from 'zod';
+import { z } from 'zod/v4';
 
 export type SimpleStatus =
   | 'no_paper'