--- a/libs/auth/src/certs.ts
+++ b/libs/auth/src/certs.ts
@@ -1,5 +1,5 @@
 import { Buffer } from 'node:buffer';
-import { z } from 'zod';
+import { z } from 'zod/v4';
 import {
   assert,
   DateWithoutTime,