--- a/libs/auth/src/cryptography.ts
+++ b/libs/auth/src/cryptography.ts
@@ -3,7 +3,7 @@ import * as fs from 'node:fs/promises';
 import path from 'node:path';
 import { Readable, Stream } from 'node:stream';
 import { FileResult, fileSync } from 'tmp';
-import { z } from 'zod';
+import { z } from 'zod/v4';
 import { unsafeParse } from '@votingworks/types';
 import { assert } from '@votingworks/basics';
 import {