--- a/libs/auth/src/keys.ts
+++ b/libs/auth/src/keys.ts
@@ -1,5 +1,5 @@
 import { Buffer } from 'node:buffer';
-import { z } from 'zod';
+import { z } from 'zod/v4';
 import { throwIllegalValue } from '@votingworks/basics';
 
 /**