--- a/libs/backend/src/language_and_audio/app_strings.ts
+++ b/libs/backend/src/language_and_audio/app_strings.ts
@@ -1,6 +1,6 @@
 import * as fs from 'node:fs/promises';
 import path from 'node:path';
-import { z } from 'zod';
+import { z } from 'zod/v4';
 import {
   MachineVersion,
   safeParseJson,