--- a/libs/backend/src/language_and_audio/vendored_translations.ts
+++ b/libs/backend/src/language_and_audio/vendored_translations.ts
@@ -1,4 +1,4 @@
-import { z } from 'zod';
+import { z } from 'zod/v4';
 import {
   safeParse,
   NonEnglishLanguageCode,