--- a/libs/backend/src/language_and_audio/vendored_translations.ts
+++ b/libs/backend/src/language_and_audio/vendored_translations.ts
@@ -16,9 +16,9 @@ export type VendoredTranslations = Record<
 >;
 
 const VendoredTranslationsSchema: z.ZodSchema<VendoredTranslations> = z.object({
-  [LanguageCode.CHINESE_SIMPLIFIED]: z.record(z.string()),
-  [LanguageCode.CHINESE_TRADITIONAL]: z.record(z.string()),
-  [LanguageCode.SPANISH]: z.record(z.string()),
+  [LanguageCode.CHINESE_SIMPLIFIED]: z.record(z.string(), z.string()),
+  [LanguageCode.CHINESE_TRADITIONAL]: z.record(z.string(), z.string()),
+  [LanguageCode.SPANISH]: z.record(z.string(), z.string()),
 });
 
 /**