--- a/libs/cdf-schema-builder/src/parse_json.ts
+++ b/libs/cdf-schema-builder/src/parse_json.ts
@@ -1,5 +1,5 @@
 import { err, ok, Result, wrapException } from '@votingworks/basics';
-import { z } from 'zod';
+import { z } from 'zod/v4';
 
 function safeParse<T>(
   parser: z.ZodType<T>,