--- a/libs/fixture-generators/src/generate-election/config.ts
+++ b/libs/fixture-generators/src/generate-election/config.ts
@@ -1,5 +1,5 @@
 /* eslint-disable vx/gts-jsdoc */
-import z from 'zod';
+import z from 'zod/v4';
 
 const positiveInt = z.number().int().min(1);