--- a/libs/fixture-generators/src/generate-election/config.ts
+++ b/libs/fixture-generators/src/generate-election/config.ts
@@ -60,3 +60,13 @@ export const defaultConfig: GenerateElectionConfig = {
     contestBallotMeasureOptionLabel: 100,
   },
 };
+
+/**
+ * Zod's `deepPartial` is not available in v4, so we manually create a
+ * schema in which all properties, including nested properties, are optional.
+ */
+export const DeepPartialGenerateElectionConfigSchema = z.object({
+  ...GenerateElectionConfigSchema.partial().shape,
+  maxStringLengths:
+    GenerateElectionConfigSchema.shape.maxStringLengths.partial(),
+});