--- a/libs/fs/src/election.ts
+++ b/libs/fs/src/election.ts
@@ -3,7 +3,7 @@ import {
   ElectionDefinition,
   safeParseElectionDefinition,
 } from '@votingworks/types';
-import { ZodError } from 'zod';
+import { ZodError } from 'zod/v4';
 import { ReadFileError, readFile } from './read_file';
 
 /**