--- a/libs/logging/src/types.ts
+++ b/libs/logging/src/types.ts
@@ -4,7 +4,7 @@ import {
   UserRole,
   UserRoleSchema,
 } from '@votingworks/types';
-import { z } from 'zod';
+import { z } from 'zod/v4';
 import {
   AppName,
   LogEventId,