--- a/libs/types/src/auth/auth.ts
+++ b/libs/types/src/auth/auth.ts
@@ -1,4 +1,4 @@
-import { z } from 'zod';
+import { z } from 'zod/v4';
 
 import { DateWithoutTime } from '@votingworks/basics';
 import { BallotStyleId, Election, ElectionId, PrecinctId } from '../election';