--- a/libs/types/src/cdf/election-event-logging/index.ts
+++ b/libs/types/src/cdf/election-event-logging/index.ts
@@ -2,7 +2,7 @@
 
 /* eslint-disable */
 
-import { z } from 'zod';
+import { z } from 'zod/v4';
 
 import check8601 from '@antongolub/iso8601';