--- a/libs/types/src/converter_client.ts
+++ b/libs/types/src/converter_client.ts
@@ -1,4 +1,4 @@
-import { z } from 'zod';
+import { z } from 'zod/v4';
 
 export type ConverterClientType = 'ms-sems';