--- a/libs/types/src/election.ts
+++ b/libs/types/src/election.ts
@@ -1,6 +1,6 @@
 import { Optional, DateWithoutTime } from '@votingworks/basics';
 import { sha256 } from 'js-sha256';
-import * as z from 'zod';
+import { z } from 'zod/v4';
 import {
   Dictionary,
   Sha256Hash,