--- a/libs/types/src/election_package_metadata.ts
+++ b/libs/types/src/election_package_metadata.ts
@@ -1,5 +1,5 @@
 /* istanbul ignore file - will eventually be tested via consumers. */
-import { z } from 'zod';
+import { z } from 'zod/v4';
 
 export type MachineVersion = 'latest';