--- a/libs/types/src/generic.ts
+++ b/libs/types/src/generic.ts
@@ -49,7 +49,7 @@ export function safeParseJson(text: string): Result<unknown, SyntaxError>;
  */
 export function safeParseJson<T>(
   text: string,
-  parser: z.ZodType<T, z.ZodTypeDef, unknown>
+  parser: z.ZodType<T>
 ): Result<T, z.ZodError | SyntaxError>;
 export function safeParseJson<T>(
   text: string,