--- a/libs/types/src/geometry.ts
+++ b/libs/types/src/geometry.ts
@@ -1,4 +1,4 @@
-import { z } from 'zod';
+import { z } from 'zod/v4';
 
 export interface Point {
   readonly x: number;