… apps/admin/backend/package.json
… apps/admin/frontend/package.json
… apps/central-scan/backend/package.json
… apps/central-scan/frontend/package.json
… apps/design/backend/package.json
… apps/design/frontend/package.json
… apps/mark-scan/backend/package.json
… apps/mark/backend/package.json
… apps/pollbook/backend/package.json
… apps/scan/backend/package.json
… apps/scan/frontend/package.json
… docs/best_practices/typescript.md
… libs/api/package.json
… libs/auth/package.json
… libs/auth/src/artifact_authentication.test.ts
… libs/backend/package.json
… libs/cdf-schema-builder/package.json
… libs/cdf-schema-builder/src/index.test.ts
… libs/cdf-schema-builder/src/parse_json.test.ts
… libs/fixture-generators/package.json
… libs/fs/package.json
… libs/logging/package.json
… libs/mark-flow-ui/package.json
… libs/pdi-scanner/package.json
… libs/printing/package.json
… libs/types/package.json
… libs/types/src/__snapshots__/election.test.ts.snap
… libs/types/src/__snapshots__/schema.test.ts.snap
… libs/types/src/__snapshots__/system_settings.test.ts.snap
… libs/types/src/generic.test.ts