## Overview

Closes #6521. Upgrades Zod to latest (package-wise, the latest version of v3) which makes v4 available. This PR then addresses the breaking changes:
- `z.record()` can no longer be called with only the property argument
- `input` field is now required in `ZodError`
- `ZodType` now only takes two parameters. The 'Input' parameter is assumed to be `unknown` rather than equal to `Output`. I think this is safe in cases where we use `ZodType` without specifying the `Input` because our code is making _fewer_ assumptions about data rather than more.
- `errors` property is now `issues`, although I couldn't find this in the changelog
- removes use of `deepPartial`

In addition, removes some things that were marked as deprecated:
- use of `superRefine`
- use of `nativeEnum`
- misc.

Needed to update tests that snapshotted `zod` error output, as the formatting has changed.

## Performance Impact

I did a little performance test, by editing a test in `./libs/types/election.test.ts` to parse an election definition 50,000 times. With v4, there was about a 50% improvement (8s to 4s) for successful parses and about a 60% improvement (15s to 6s) for error parses. 

I then experimented with a stripped down version of the "happy path - election flow" test in `./apps/admin/backend/app.cvr_files.test.ts`, importing and clearing the mock CVR file 20 times. There was no difference between `main` and this branch, both took 22 seconds. There could be a meaningful difference in a real world setting with larger files or on weaker CPUs (e.g. VxMark), but I suspect that the parsing takes up too small of a portion of operations, outside the logging use case, to make a big difference.

## Other Notes

The first time I tried upgrading `zod`, I used `pnpm -r up zod@latest`. This worked, but it changed some transitive dependency that made a bunch of React Testing Library assertions fail without updating things to be more async, like `.getByText` to `.findByText`. This is actually not the first time I've experienced this during an upgrade. I think it's some sort of transitive dependency change. Anyone else seen this?